tests: create test case
